+++ java-manta-it/src/test/java/com/joyent/manta/client/MantaJobBuilderIT.java
@@ -1,0 +1,79 @@
+/**
+ * Copyright (c) 2015, Joyent, Inc. All rights reserved.
+ */
+package com.joyent.manta.client;
+
+import com.joyent.manta.client.config.TestConfigContext;
+import com.joyent.manta.config.ConfigContext;
+import com.joyent.manta.exception.MantaCryptoException;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.Optional;
+import org.testng.annotations.Parameters;
+import org.testng.annotations.Test;
+
+import java.io.IOException;
+import java.util.List;
+import java.util.UUID;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
+
+/**
+ * @author <a href="https://github.com/dekobon">Elijah Zupancic</a>
+ */
+public class MantaJobBuilderIT {
+    private static final String TEST_DATA = "EPISODEII_IS_BEST_EPISODE";
+
+    private String testPathPrefix;
+
+    private AtomicInteger count = new AtomicInteger(0);
+
+    private MantaClient mantaClient;
+
+    @BeforeClass
+    @Parameters({"manta.url", "manta.user", "manta.key_path", "manta.key_id", "manta.timeout"})
+    public void beforeClass(@Optional String mantaUrl,
+                            @Optional String mantaUser,
+                            @Optional String mantaKeyPath,
+                            @Optional String mantaKeyId,
+                            @Optional Integer mantaTimeout)
+            throws IOException, MantaCryptoException {
+
+        // Let TestNG configuration take precedence over environment variables
+        ConfigContext config = new TestConfigContext(
+                mantaUrl, mantaUser, mantaKeyPath, mantaKeyId, mantaTimeout);
+
+        mantaClient = new MantaClient(config);
+
+        testPathPrefix = String.format("%s/stor/%s",
+                config.getMantaHomeDirectory(), UUID.randomUUID());
+        mantaClient.putDirectory(testPathPrefix, null);
+    }
+
+
+    @AfterClass
+    public void cleanup() throws IOException {
+        if (mantaClient != null) {
+            mantaClient.deleteRecursive(testPathPrefix);
+            mantaClient.closeQuietly();
+        }
+    }
+
+    @Test
+    public void canBuildTypicalJob() throws IOException {
+        final MantaJobBuilder builder = mantaClient.jobBuilder();
+
+        MantaJobBuilder.Done finishedJob = builder.newJob("foo")
+               .addPhase(new MantaJobPhase()
+                            .setExec("")
+                            .setType("map"))
+               .addInput("")
+               .run()
+               .waitUntilDone();
+
+        List<String> outputs = finishedJob
+                .throwExceptionIfFailed()
+                .outputs()
+                .collect(Collectors.toList());
+    }
+}